+++ document/src/main/java/com/yahoo/document/json/document/DocumentParser.java
@@ -4,17 +4,24 @@ package com.yahoo.document.json.document;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.yahoo.document.DocumentId;
+import com.yahoo.document.json.JsonReader;
 import com.yahoo.document.json.readers.DocumentParseInfo;
 
 import java.io.IOException;
 import java.util.Optional;
 
-import static com.yahoo.document.json.JsonReader.*;
+import static com.yahoo.document.json.JsonReader.bufferFields;
+import static com.yahoo.document.json.JsonReader.nextToken;
+import static com.yahoo.document.json.readers.JsonParserHelpers.expectObjectStart;
 
 public class DocumentParser {
     private static final String UPDATE = "update";
     private static final String PUT = "put";
     private static final String ID = "id";
+    private static final String CONDITION = "condition";
+    public static final String CREATE_IF_NON_EXISTENT = "create";
+    public static final String FIELDS = "fields";
+    public static final String REMOVE = "remove";
 
     public static Optional<DocumentParseInfo> parseDocument(JsonParser parser) {
         // we should now be at the start of a feed operation or at the end of the feed
@@ -68,19 +75,70 @@ public class DocumentParser {
         }
     }
 
-    private static SupportedOperation operationNameToOperationType(String operationName) {
+    private static JsonReader.SupportedOperation operationNameToOperationType(String operationName) {
         switch (operationName) {
             case PUT:
             case ID:
-                return SupportedOperation.PUT;
+                return JsonReader.SupportedOperation.PUT;
             case REMOVE:
-                return SupportedOperation.REMOVE;
+                return JsonReader.SupportedOperation.REMOVE;
             case UPDATE:
-                return SupportedOperation.UPDATE;
+                return JsonReader.SupportedOperation.UPDATE;
             default:
                 throw new IllegalArgumentException(
                         "Got " + operationName + " as document operation, only \"put\", " +
                                 "\"remove\" and \"update\" are supported.");
         }
+    }
+
+
+    public static DocumentParseInfo parseToDocumentsFieldsAndInsertFieldsIntoBuffer(JsonParser parser, DocumentId documentId) {
+        long indentLevel = 0;
+        DocumentParseInfo documentParseInfo = new DocumentParseInfo();
+        documentParseInfo.documentId = documentId;
+        while (true) {
+            // we should now be at the start of a feed operation or at the end of the feed
+            JsonToken t = nextToken(parser);
+            if (t == null) {
+                throw new IllegalArgumentException("Could not read document, no document?");
+            }
+            switch (t) {
+                case START_OBJECT:
+                    indentLevel++;
+                    break;
+                case END_OBJECT:
+                    indentLevel--;
+                    break;
+                case START_ARRAY:
+                    indentLevel+=10000L;
+                    break;
+                case END_ARRAY:
+                    indentLevel-=10000L;
+                    break;
+            }
+            if (indentLevel == 1 && (t == JsonToken.VALUE_TRUE || t == JsonToken.VALUE_FALSE)) {
+                try {
+                    if (CREATE_IF_NON_EXISTENT.equals(parser.getCurrentName())) {
+                        documentParseInfo.create = Optional.ofNullable(parser.getBooleanValue());
+                        continue;
+                    }
+                } catch (IOException e) {
+                    throw new RuntimeException("Got IO exception while parsing document", e);
+                }
+            }
+            if (indentLevel == 2L && t == JsonToken.START_OBJECT) {
+
+                try {
+                    if (!FIELDS.equals(parser.getCurrentName())) {
+                        continue;
+                    }
+                } catch (IOException e) {
+                    throw new RuntimeException("Got IO exception while parsing document", e);
+                }
+                bufferFields(parser, documentParseInfo.fieldsBuffer, t);
+                break;
+            }
+        }
+        return documentParseInfo;
     }
 }
