+++ document/src/main/java/com/yahoo/document/json/readers/SingleValueReader.java
@@ -3,10 +3,13 @@ package com.yahoo.document.json.readers;
 
 import com.fasterxml.jackson.core.JsonToken;
 import com.yahoo.document.DataType;
+import com.yahoo.document.DocumentId;
+import com.yahoo.document.PositionDataType;
+import com.yahoo.document.ReferenceDataType;
 import com.yahoo.document.datatypes.FieldValue;
-import com.yahoo.document.json.JsonReader;
 import com.yahoo.document.json.TokenBuffer;
 import com.yahoo.document.update.ValueUpdate;
+import org.apache.commons.codec.binary.Base64;
 
 public class SingleValueReader {
     public static final String UPDATE_ASSIGN = "assign";
@@ -17,7 +20,7 @@ public class SingleValueReader {
 
     public static FieldValue readSingleValue(TokenBuffer buffer, JsonToken t, DataType expectedType) {
         if (t.isScalarValue()) {
-            return JsonReader.readAtomic(buffer, expectedType);
+            return readAtomic(buffer, expectedType);
         } else {
             return CompositeReader.createComposite(buffer, expectedType);
         }
@@ -52,4 +55,24 @@ public class SingleValueReader {
         return update;
     }
 
+    public static FieldValue readAtomic(TokenBuffer buffer, DataType expectedType) {
+        if (expectedType.equals(DataType.RAW)) {
+            return expectedType.createFieldValue(new Base64().decode(buffer.currentText()));
+        } else if (expectedType.equals(PositionDataType.INSTANCE)) {
+            return PositionDataType.fromString(buffer.currentText());
+        } else if (expectedType instanceof ReferenceDataType) {
+            return readReferenceFieldValue(buffer, expectedType);
+        } else {
+            return expectedType.createFieldValue(buffer.currentText());
+        }
+    }
+
+    private static FieldValue readReferenceFieldValue(TokenBuffer buffer, DataType expectedType) {
+        final FieldValue value = expectedType.createFieldValue();
+        final String refText = buffer.currentText();
+        if (!refText.isEmpty()) {
+            value.assign(new DocumentId(buffer.currentText()));
+        }
+        return value;
+    }
 }
