+++ document/src/main/java/com/yahoo/document/json/readers/TensorReader.java
@@ -2,10 +2,11 @@
 package com.yahoo.document.json.readers;
 
 import com.yahoo.document.datatypes.TensorFieldValue;
-import com.yahoo.document.json.JsonReader;
 import com.yahoo.document.json.TokenBuffer;
 import com.yahoo.tensor.MappedTensor;
 import com.yahoo.tensor.Tensor;
+
+import static com.yahoo.document.json.readers.JsonParserHelpers.*;
 
 public class TensorReader {
     public static final String TENSOR_ADDRESS = "address";
@@ -15,27 +16,27 @@ public class TensorReader {
 
     public static void fillTensor(TokenBuffer buffer, TensorFieldValue tensorFieldValue) {
         Tensor.Builder tensorBuilder = Tensor.Builder.of(tensorFieldValue.getDataType().getTensorType());
-        JsonReader.expectObjectStart(buffer.currentToken());
+        expectObjectStart(buffer.currentToken());
         int initNesting = buffer.nesting();
-        // read tensor cell fields and ignore everything else
+        /* read tensor cell fields and ignore everything else */
         for (buffer.next(); buffer.nesting() >= initNesting; buffer.next()) {
             if (TensorReader.TENSOR_CELLS.equals(buffer.currentName()))
                 readTensorCells(buffer, tensorBuilder);
         }
-        JsonReader.expectObjectEnd(buffer.currentToken());
+        expectObjectEnd(buffer.currentToken());
         tensorFieldValue.assign(tensorBuilder.build());
     }
 
     public static void readTensorCells(TokenBuffer buffer, Tensor.Builder tensorBuilder) {
-        JsonReader.expectArrayStart(buffer.currentToken());
+        expectArrayStart(buffer.currentToken());
         int initNesting = buffer.nesting();
         for (buffer.next(); buffer.nesting() >= initNesting; buffer.next())
             readTensorCell(buffer, tensorBuilder);
-        JsonReader.expectCompositeEnd(buffer.currentToken());
+        expectCompositeEnd(buffer.currentToken());
     }
 
     public static void readTensorCell(TokenBuffer buffer, Tensor.Builder tensorBuilder) {
-        JsonReader.expectObjectStart(buffer.currentToken());
+        expectObjectStart(buffer.currentToken());
         int initNesting = buffer.nesting();
         double cellValue = 0.0;
         Tensor.Builder.CellBuilder cellBuilder = tensorBuilder.cell();
@@ -47,18 +48,18 @@ public class TensorReader {
                 cellValue = Double.valueOf(buffer.currentText());
             }
         }
-        JsonReader.expectObjectEnd(buffer.currentToken());
+        expectObjectEnd(buffer.currentToken());
         cellBuilder.value(cellValue);
     }
 
     public static void readTensorAddress(TokenBuffer buffer, MappedTensor.Builder.CellBuilder cellBuilder) {
-        JsonReader.expectObjectStart(buffer.currentToken());
+        expectObjectStart(buffer.currentToken());
         int initNesting = buffer.nesting();
         for (buffer.next(); buffer.nesting() >= initNesting; buffer.next()) {
             String dimension = buffer.currentName();
             String label = buffer.currentText();
             cellBuilder.label(dimension, label);
         }
-        JsonReader.expectObjectEnd(buffer.currentToken());
+        expectObjectEnd(buffer.currentToken());
     }
 }
